// gigmosaic-fe/src/core/models/interface.tsx
export interface AdditionalService {
  id: number;
  name: string;
  price: number;
  duration: string;
  images: FileList | null;
}

type SeoItem = {
  id: string;
  metaDescription: string;
  metaKeywords: string[];
  metaTitle: string;
};

export interface ServiceUpdate {
  serviceId: string;
  serviceTitle: string;
  slug: string;
  categoryId: string;
  serviceProvider: string;
  subCategoryId: string;
  duration: number;
  serviceOverview: string;
  amount: number;
  price: number;
  priceAfterDiscount: number;
  offerType?: string;
  offerAmount?: number;
  offerDurationType?: string;
  offerDuration?: string;
  afterDiscountAmount?: number;
  staff: string[];
  includes?: string[];
  isActive: boolean;
  additionalService?: AdditionalService[];
  isOffers?: boolean;
  gallery: [
    {
      serviceImages: [];
      videoLink: string;
    },
  ];
  location: [
    {
      address: string;
      city: string;
      state: string;
      country: string;
      pinCode: string;
      googleMapsPlaceId: string;
      coordinates: {
        longitude: number;
        latitude: number;
      };
    },
  ];
  additionalServices: [{
    images: string;
    price: number;
    serviceItem: string;
  }];
  faq: [{
    question: string;
    answer: string;
  }];
  seo: SeoItem[];
  availability: [{
    alldate: boolean;
    day:string;
    timeSlots:[{
      from:string;
      to:string
    }]
    Monday?: { from: string; to: string }[];
    Tuesday?: { from: string; to: string }[];
    Wednesday?: { from: string; to: string }[];
    Thursday?: { from: string; to: string }[];
    Friday?: { from: string; to: string }[];
    Saturday?: { from: string; to: string }[];
    Sunday?: { from: string; to: string }[];
  }]
}